src/backend/requirements.txt
# 后端依赖
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
requests==2.31.0
pyyaml==6.0.1
ffmpeg-python==0.2.0
python-docx==1.1.0
